+++ CHANGELOG.md
@@ -41,6 +41,7 @@ All `0.1.0-alpha` releases are unstable, and have missing widgets.
   - `/blacklist <login>` was renamed to `/blacklist add <login>`
   - `/unblacklist <login>` was renamed to `/blacklist remove <login>`
   - Added `/blacklist clear` which removes all players from the blacklist.
+  - `/map_import <id/uid>` was renamed to `/import map <id/uid>`
   
 ### [Commits](https://github.com/timwie/steward/compare/v0.1.0-alpha5...v0.1.0-alpha6)